requirement.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas
streamlit
pillow
scikit-learn
seaborn